add screenshot function and cleanup

+++ include/splatc/ppm.h
@@ -3,6 +3,6 @@
 #include <stdint.h>
 #include <stdlib.h>
 
-void ppm_write(uint8_t* pixels, size_t width, size_t height, char* fn);
+void ppm_write(float* pixels, size_t width, size_t height, char* fn);
 
 #endif
